create a download module
